notify systemd after endpoints are set up
The USB gadget needs to be activated after uMTP has set up the endpoints.
It is currently not possible to tell from the outside whether this has
already happened or not. This is usually circumvented by a workaround such
as a fixed waiting time. On platforms using systemd this can be neatly
solved with a notify service (type=notify). To take advantage of this, an
option is added to notify systemd once the endpoint setup is complete.

Signed-off-by: Matthias Fend <matthias.fend@emfend.at>