Make PRINT_* macros behave consistently
The function-like PRINT_* macros would build differently for different
build options, such that something like:

  if (condition)
    PRINT_MSG("yes")
  else
    PRINT_MSG("no")

(note the missing semicolons)
would build successfully with no additional build flags or with DEBUG=1,
but would fail to compile with USE_SYSLOG=1.

This change makes it so that all build options cause the above to fail,
same as if it had been a regular function call.