mtp_cfg: add option to set the MTP_Extensions string
Add a configuration option to allow a user to set the mtp extensions
the responder will advertise to the host. This is only provided for
backwards compatibility.

Note that this string is defaulting to empty, and the string inherited
from the tizen code-base mentioned in the example configuration can be
used but can&will cause a hosts MTP implementation to see extensions
that are not backed by any logic - and in the case of libmtp will
trigger vendor specific quirk handling that unloads the hosts usb
driver and resets the usb device.

Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>